phase3_ai-chatbot/chatbot_backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
openai
python-dotenv
